added annotation traceable with cpu time and memory time

+++ src/main/java/com/example/testdatalogclient/config/CustomHttpInterceptor.java
@@ -50,7 +50,7 @@ public class CustomHttpInterceptor implements HandlerInterceptor {
             }
             if(t.memoryUsage()){
                 OperatingSystemMXBean osBean = (OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
-                double memoryUsage = osBean.getTotalMemorySize();
+                long memoryUsage = osBean.getTotalMemorySize();
                 execTime.setMemoryUsage(memoryUsage);
 
             }
